requirements.txt
Django==4.0.3
django-cors-headers==3.13.0
requests==2.26.0
python-whois==0.8.0
celery==5.2.7
amqp==5.1.1
gevent==22.10.2
eventlet==0.33.2
SQLAlchemy==1.4.40
psycopg2-binary==2.9.5
six==1.14.0
coreapi==2.3.3
fcm-django==1.0.12
firebase-admin==5.4.0